+++ alluxio/BUILD
@@ -1,0 +1,14 @@
+package(default_visibility = ["//visibility:public"])
+
+py_test(
+    name = "test_alluxio",
+    size = "enormous",
+    srcs = ["test_alluxio.py"],
+    data = ["alluxio.sh"],
+    local = True,
+    shard_count = 1,
+    deps = [
+        "//integration_tests:dataproc_test_case",
+        "@io_abseil_py//absl/testing:parameterized",
+    ],
+)
