+++ hbase/test_hbase.py
@@ -66,4 +66,3 @@ class HBaseTestCase(DataprocTestCase):
 
 if __name__ == '__main__':
     absltest.main()
-
